bot/requirements.txt
requests
pynput
pyperclip
beautifulSoup4
psutil
paramiko
pillow
pycryptodome